enhanced-research-agent/requirements.txt
# Core dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Web search and browsing
duckduckgo-search>=3.8.0
requests>=2.28.0
beautifulsoup4>=4.12.0

# Data processing and visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0

# Utility libraries
python-dateutil>=2.8.2
